durable-image-metadata/requirements.txt
azure-functions
Pillow
azure-functions-durable
